Fix missing model diagram in end-user version
Add full Graphviz package to UniSim installation file
Include Graphviz in fill_folders.bat

+++ src/plugins/awe_models/weather.cpp
@@ -44,7 +44,7 @@ void Weather::reset()
 
 void Weather::update()
 {
-    int dayOfYear = int(calendar->pullVariable<double>("dayOfYear"));
+    int dayOfYear = calendar->pullVariable<int>("dayOfYear");
     T = temperature(dayOfYear);
     Tsum = (dayOfYear == 1) ? T : Tsum + T;
 }
